# docs/static/swagger/openApiConfig.yaml
openapi: 3.0.3
info:
  title: ''
  version: ''
  description: >
    This is a simple CRUD API application made with Express and documented with
    Swagger
  termsOfService: http://swagger.io/terms/
  contact:
    email: bourdakos1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-github: https://github.com/bee-travels
servers:
  - url: http://localhost:3001/api
    description: Development server server (uses local data)
  - url: https://sandbox-api.example.com:8443/v1
    description: Sandbox server (uses test data)
...
paths:
...
  /drivers:
    get:
      tags:
        - Drivers
      summary: Get a list of drivers
      description: Returns a list of drivers along with their user and vehicle information.
      operationId: getDrivers
      responses:
        '200':
          description: Successful operation, returns a list of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
        '400':
          description: Error occurred while obtaining the driver list
  /drivers/business/:business_id:
    get:
      tags:
        - Drivers
      summary: Get a list of drivers for business
      description: Returns a list of drivers along with their user and vehicle information.
      operationId: getDrivers
      responses:
        '200':
          description: Successful operation, returns a list of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
        '400':
          description: Error occurred while obtaining the driver list
  /drivers/full:
    get:
      tags:
        - Drivers
      summary: Get a list of drivers
      description: Returns a list of drivers along with their user and vehicle information.
      operationId: getDrivers
      responses:
        '200':
          description: Successful operation, returns a list of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
        '400':
          description: Error occurred while obtaining the driver list
...
  /vehicles:
    get:
      tags:
        - Vehicles
      summary: List all vehicles
      description: Retrieves a list of all vehicles in the database.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '500':
          description: Error fetching vehicles
    patch:
      tags:
        - Vehicles
      summary: Update a vehicle
      description: Updates an existing vehicle's details and specifications.
      operationId: updateVehicle
      parameters:
        - name: vehicle_id
          in: path
          description: The ID of the vehicle to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdate'
        required: true
      responses:
        '200':
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Error updating vehicle
...
  /vehicles/:vehicle_id:
    get:
      tags:
        - Vehicles
      summary: Get a vehicle by ID
      description: Retrieves a single vehicle by its ID from the database.
      parameters:
        - name: vehicle_id
          in: path
          description: The ID of the vehicle to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          description: Vehicle not found
        '500':
          description: Error fetching vehicle
    delete:
      tags:
        - Vehicles
      summary: Delete a vehicle
      description: Deletes a vehicle from the database.
      operationId: deleteVehicle
      parameters:
        - name: vehicle_id
          in: path
          description: The ID of the vehicle to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vehicle deleted successfully
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Error deleting vehicle
...
  /vehicles/create:
    post:
      tags:
        - Vehicles
      summary: Create a new vehicle
      description: Adds a new vehicle to the database, including its specifications.
      operationId: createNewVehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
        required: true
      responses:
        '201':
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Error creating vehicle
...